+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### dev
+- Add Spotlight Omnisearch integration
+
 ### v1.3.0
 - Add QuickInsert integration
 - Visible lock on module-loaded spelllists which can't be disabled